+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - 
 ### Changed:
 - datapack tag system
+- max metal count to 7